refactor: dukan categories ui state
