+++ dukan_presentation/src/commonMain/kotlin/net/thechance/mena/dukan/presentation/viewModel/dukanCategories/DukanCategoriesInteractionListener.kt
@@ -2,6 +2,6 @@ package net.thechance.mena.dukan.presentation.viewModel.dukanCategories
 
 interface DukanCategoriesInteractionListener {
     fun onBackClicked()
-    fun onCategoryClicked(categoryName:String,categoryId: String)
+    fun onCategoryClicked(categoryName: String, categoryId: String)
     fun onDismissSnackBar()
 }
