+++ dukan_presentation/src/commonMain/kotlin/net/thechance/mena/dukan/presentation/viewModel/dukanCategories/DukanCategoriesUiState.kt
@@ -5,10 +5,10 @@ import net.thechance.mena.dukan.presentation.component.SnackBarUiState
 data class DukanCategoriesUiState(
     val snackBarUiState: SnackBarUiState?=null,
     val categories: List<CategoryUiState> = emptyList(),
-)
-
-data class CategoryUiState(
-    val id: String,
-    val name: String,
-    val imageUrl: String,
-)
+) {
+    data class CategoryUiState(
+        val id: String,
+        val name: String,
+        val imageUrl: String,
+    )
+}
