+++ dukan_presentation/src/commonMain/kotlin/net/thechance/mena/dukan/presentation/viewModel/dukanCategories/Mapper.kt
@@ -2,8 +2,8 @@ package net.thechance.mena.dukan.presentation.viewModel.dukanCategories
 
 import net.thechance.mena.dukan.domain.entity.Category
 
-fun Category.toUiState(): CategoryUiState {
-    return CategoryUiState(
+fun Category.toUiState(): DukanCategoriesUiState.CategoryUiState {
+    return DukanCategoriesUiState.CategoryUiState(
         id = id,
         name = name,
         imageUrl = imageUrl
